drastically reorganize all http-specifics into otter/http module
