+++ examples/01-hello-world/src/http/greetings/greeting-model.ts
