+++ examples/01-hello-world/src/index.ts
@@ -1,5 +1,7 @@
 import {otter} from 'otter';
 
 otter({
-  port: 8080,
+  http: {
+    port: 8080
+  },
 });
