--- src/controller/coerce/controller.ts
@@ -1,6 +1,0 @@
-import {Controller} from 'src/controller/types';
-import {WebController} from 'src/controller';
-
-export function controller(controller: Controller): WebController {
-  return new WebController(controller);
-}
