--- src/controller/index.ts
@@ -1,2 +1,0 @@
-export * from './coerce';
-export * from './web-controller';
