--- src/controller/types/controller.ts
@@ -1,12 +1,0 @@
-import {ActionPipeline, HttpInput} from 'src/action-pipeline';
-import {RestRoute} from './rest-route';
-
-export type HttpBindingInfo = {
-  type: 'http';
-  route: RestRoute<any, any>;
-}
-
-export type Controller = {
-  actions: Array<ActionPipeline<HttpInput, any, HttpBindingInfo>>
-}
-
