--- src/controller/types/index.ts
@@ -1,3 +1,0 @@
-export * from './action'
-export * from './controller';
-export * from './rest-route';
