--- src/controller/web-controller/index.ts
@@ -1,1 +1,0 @@
-export * from './web-controller';
