+++ src/http/coerce/index.ts
@@ -1,4 +1,4 @@
 export * from './action';
 export * from './controller';
 export * from './extract';
-export * from './rest';
+export * from './route';
