+++ src/http/http-binding-info.ts
@@ -1,0 +1,10 @@
+import {HttpMethod} from '../types';
+
+export type HttpBindingInfo = {
+  type: 'http';
+  route: {
+    method: HttpMethod,
+    path: string;
+  };
+}
+
