+++ src/http/startup/http-options.ts
@@ -1,0 +1,3 @@
+export type HttpOptions = {
+  port: number;
+}
