+++ src/index.ts
@@ -1,5 +1,1 @@
-export {action, controller, extract, rest} from './controller';
-export {HttpError} from './error'
-export {Extract} from './extract';
 export {otter} from './otter';
-export {Schema} from './schema'
