+++ src/otter/otter.ts
@@ -1,27 +1,15 @@
-import * as express from 'express';
-import {json} from 'body-parser'
-import {discoverControllers} from './discover-controllers';
+import {HttpOptions, startHttpServer} from 'src/http/startup';
 
 type OtterOptions = {
-  port: number
+  http?: HttpOptions
 }
 
 async function initialize(options: OtterOptions): Promise<void> {
-  const app = express();
-  app.use(json());
-
-  await discoverControllers(app);
-
-  return new Promise((resolve, reject) => {
-    app.listen(options.port, resolve);
-  });
+  if (options.http) await startHttpServer(options.http);
 }
 
 export function otter(options: OtterOptions) {
   initialize(options)
-    .then(() => {
-      console.log(`[otter] Initialized and listening on port ${options.port}`)
-    })
     .catch(e => {
       throw e
     });
